requirements.txt
flask[async]
python-telegram-bot==20.8
openai>=1.0.0
hypercorn